requirements.txt
pyglet
pillow
numpy